Added admin and user routers with role attached in JWT Token

+++ Projects/Project3/Todo_App/router/todos.py
@@ -40,14 +40,14 @@ class TodoRequest(BaseModel):
 async def get_all_records(user:user_auth,db:db_dependency):
     if user is None:
         raise HTTPException(status_code=401,detail="Authentication Failed")
-    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
+    return db.query(Todos).filter(Todos.owner_id == user.get('user_id')).all()
 
 # get todo by id
 @router.get("/records/{todo_id}",status_code=status.HTTP_200_OK)
 async def get_todo_by_id(user:user_auth,db:db_dependency,todo_id:int = Path(gt=0)):
     if user is None:
         raise HTTPException(status_code=401,detail="Authentication Failed")
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('user_id')).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404,detail="Todo ID not found")
@@ -57,14 +57,16 @@ async def get_todo_by_id(user:user_auth,db:db_dependency,todo_id:int = Path(gt=0
 async def create_todo(user:user_auth,db:db_dependency,todo_request:TodoRequest):
     if user is None:
         raise HTTPException(status_code=401,detail="Authentication Failed")
-    todo_model = Todos(**todo_request.model_dump(),owner_id = user.get('id'))
+    todo_model = Todos(**todo_request.model_dump(),owner_id = user.get('user_id'))
     db.add(todo_model)
     db.commit()
 
 # update an existing todo
 @router.put("/update_todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db:db_dependency,todo_updated:TodoRequest,todo_id:int=Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user:user_auth,db:db_dependency,todo_updated:TodoRequest,todo_id:int=Path(gt=0)):
+    if user is None:
+            raise HTTPException(status_code=401,detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('user_id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404,detail="Todo ID not found")
     # update the old with new data
@@ -78,9 +80,12 @@ async def update_todo(db:db_dependency,todo_updated:TodoRequest,todo_id:int=Path
 
 # delete a todo
 @router.delete("/delete_todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db:db_dependency,todo_id:int=Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user:user_auth,db:db_dependency,todo_id:int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("user_id")).first()
     if todo_model is None:
         raise HTTPException(status_code=404,detail="Todo not found")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("user_id")).delete()
     db.commit()
